working on fixing endpoints for serach
