+++ components/carousel/Carousel.tsx
@@ -5,54 +5,55 @@ import { BookPreview } from "../bookPreview/BookPreview";
 import { BiChevronRight, BiChevronLeft } from "react-icons/bi";
 
 interface CarouselProps {
-    books: Book[];
+	books: Book[];
 }
 
 export default function Carousel({ books }: CarouselProps) {
-    if (!books || books.length === 0) {
-        return (
-            <img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTFgCFmiB0pQYHGDklq0EZX35hq-EuOP7N8Xg&usqp=CAU" />
-        );
-    } else
-        return (
-            <>
-                <Splide
-                    hasTrack={false}
-                    aria-label="..."
-                    options={{
-                        perPage: 3,
-                        perMove: 1,
-                        focus: 0,
-                        gap: "100px",
-                        pagination: false,
-                        trimSpace: false,
-                        type: "loop",
-                    }}
-                >
-                    <SplideTrack className="py-4">
-                        {books.map((book, index) => (
-                            <SplideSlide key={index}>
-                                <BookPreview
-                                    imgSrc={book.image}
-                                    bookTitle={book.title}
-                                    bookAuthor={book.author}
-                                    linkHref={`/book/${book.identifier}`}
-                                    bookSize={"carouselItem"}
-                                />
-                            </SplideSlide>
-                        ))}
-                    </SplideTrack>
+	if (!books || books.length === 0) {
+		return (
+			<img src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTFgCFmiB0pQYHGDklq0EZX35hq-EuOP7N8Xg&usqp=CAU" />
+		);
+	} else
+		return (
+			<>
+				<Splide
+					hasTrack={false}
+					aria-label="..."
+					options={{
+						perPage: 3,
+						perMove: 1,
+						focus: 0,
+						gap: "100px",
+						pagination: false,
+						trimSpace: false,
+						type: "loop",
+					}}
+				>
+					<SplideTrack className="py-4">
+						{books.map((book, index) => (
+							<SplideSlide key={index}>
+								<BookPreview
+									isAvailable={book.isAvailable}
+									imgSrc={book.image}
+									bookTitle={book.title}
+									bookAuthor={book.author}
+									linkHref={`/book/${book.identifier}`}
+									bookSize={"carouselItem"}
+								/>
+							</SplideSlide>
+						))}
+					</SplideTrack>
 
-                    <div className="splide__arrows">
-                        <button className="splide__arrow splide__arrow--prev">
-                            <BiChevronRight className="h-full w-full " />
-                        </button>
-                        <button className="splide__arrow splide__arrow--next">
-                            <BiChevronRight className="h-full w-full" />
-                        </button>
-                    </div>
-                </Splide>
-                <div className=" pb-8 " />
-            </>
-        );
+					<div className="splide__arrows">
+						<button className="splide__arrow splide__arrow--prev">
+							<BiChevronRight className="h-full w-full " />
+						</button>
+						<button className="splide__arrow splide__arrow--next">
+							<BiChevronRight className="h-full w-full" />
+						</button>
+					</div>
+				</Splide>
+				<div className=" pb-8 " />
+			</>
+		);
 }
