+++ hooks/useGetBooks.tsx
@@ -2,20 +2,24 @@ import { Book } from "@prisma/client";
 import { useQuery } from "@tanstack/react-query";
 
 type UseGetBooksProps = {
-    category?: string;
-    orderBy?: "createdAt";
+	category?: string;
+	orderBy?: "createdAt";
+	availability?: boolean;
 };
 
-export default function useGetBooks({ category, orderBy }: UseGetBooksProps) {
-    return useQuery<Book[]>(["books"], () =>
-        fetch(
-            `http://localhost:3002/api/book?category=${category}&orderBy=${orderBy}`,
-            {
-                method: "GET",
-            }
-        ).then((res) => {
-            console.log(res.body);
-            return res.json();
-        })
-    );
+export default function useGetBooks({
+	category,
+	orderBy,
+	availability,
+}: UseGetBooksProps) {
+	return useQuery<Book[]>(["books"], () =>
+		fetch(
+			`http://localhost:3000/api/book?isAvailable=${availability}&category=${category}&orderBy=${orderBy}`,
+			{
+				method: "GET",
+			}
+		).then((res) => {
+			return res.json();
+		})
+	);
 }
