+++ pages/api/book/model.prisma
@@ -23,7 +23,7 @@ model Book {
     publishYear     DateTime?
     genres          Json
     tags            Json
-    isAvailable     Boolean @default(false)
+    isAvailable     Boolean @default(true)
     isReserved      Boolean @default(false)
     ownerId  String
     owner           User? @relation("owning", fields: [ownerId], references: [identifier])
