+++ pages/favorites.page.tsx
@@ -1,10 +1,10 @@
-
 import { Book } from "@prisma/client";
 import { BookGrid } from "../components/bookGrid/BookGrid";
-import { useBooks } from "../hooks/useBooks";
+import useGetBooks from "../hooks/useGetBooks";
 
 export default function Favorites() {
-	const { data: books, isLoading: booksLoading } = useBooks();
+	const { data: books, isLoading: booksLoading, isError } = useGetBooks({});
+	console.log(books);
 
 	if (booksLoading) return <p>Loading...</p>;
 
@@ -12,7 +12,7 @@ export default function Favorites() {
 	console.log(books);
 	return (
 		<>
-			<h2 className="border-b border-grey pb-4 pt-7 text-center font-arnoPro text-2xl font-bold text-dustyRose">
+			<h2 className="font-arnoPro border-b border-grey pb-4 pt-7 text-center text-2xl font-bold text-dustyRose">
 				Favorites
 			</h2>
 			<BookGrid books={books} booksLoading={booksLoading} />
